some sort of streaming implemented, might need fixes depending on how the front end will call the api.
