+++ src/server/trpc/router/_app.ts
@@ -2,12 +2,11 @@
 import { router } from "../trpc";
 import { exampleRouter } from "./example";
 import { authRouter } from "./auth";
-import { videoRouter } from "./video";
+import { videoRouter } from "./deletethisaftertesting";
 
 export const appRouter = router({
   example: exampleRouter,
   auth: authRouter,
-  video: videoRouter,
 });
 
 // export type definition of API
